python/requirements.txt
ecdsa==0.18.0
pycryptodome==3.17
six==1.16.0
